LayoutTests/platform/ios-simulator/ios/fast/parser/block-nesting-cap-iphone-expected.txt

Test that the HTML parser does not allow the nesting depth of "block-level" elements to exceed 768.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL d768.parentNode === d766 should be true. Was false.
FAIL d768.parentNode === d766 should be true. Was false.
PASS d768.previousSibling === d767 is true
PASS successfullyParsed is true

TEST COMPLETE
